OAK-8962 Refactoring some of the movement
